Moving checkuser to navbar
